delete tabs and generate welcome page
