+++ src/app/app-routing.module.ts
@@ -2,7 +2,8 @@ import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 
 const routes: Routes = [
-  { path: '', loadChildren: './tabs/tabs.module#TabsPageModule' }
+  { path: '', pathMatch: 'full', redirectTo: 'welcome' },
+  { path: 'welcome', loadChildren: './pages/welcome/welcome.module#WelcomePageModule' }
 ];
 @NgModule({
   imports: [
@@ -10,4 +11,4 @@ const routes: Routes = [
   ],
   exports: [RouterModule]
 })
-export class AppRoutingModule {}
+export class AppRoutingModule { }
