+++ src/app/pages/welcome/welcome.module.ts
@@ -1,0 +1,26 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { Routes, RouterModule } from '@angular/router';
+
+import { IonicModule } from '@ionic/angular';
+
+import { WelcomePage } from './welcome.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: WelcomePage
+  }
+];
+
+@NgModule({
+  imports: [
+    CommonModule,
+    FormsModule,
+    IonicModule,
+    RouterModule.forChild(routes)
+  ],
+  declarations: [WelcomePage]
+})
+export class WelcomePageModule {}
