+++ src/app/pages/welcome/welcome.page.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-welcome',
+  templateUrl: './welcome.page.html',
+  styleUrls: ['./welcome.page.scss'],
+})
+export class WelcomePage implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
