--- src/app/tab1/tab1.page.scss
@@ -1,4 +1,0 @@
-.welcome-card img {
-  max-height: 35vh;
-  overflow: hidden;
-}
