--- src/app/tab2/tab2.page.scss
@@ -1,1 +1,0 @@
-
