--- src/app/tab3/tab3.page.scss
@@ -1,1 +1,0 @@
-
