--- src/app/tabs/tabs.page.scss
@@ -1,1 +1,0 @@
-
